--- app/serializers/list_serializer.rb
@@ -1,7 +1,0 @@
-class ListSerializer < ApplicationSerializer
-  attributes :id, :name, :description, :published_at, :resources_count
-
-  def resources_count
-    object.resources.count
-  end
-end
