+++ app/serializers/v1/application_serializer.rb
@@ -1,0 +1,7 @@
+module V1
+  class ApplicationSerializer < ActiveModel::Serializer
+    link :self do
+      href send("api_v1_#{object.class.to_s.downcase}_url", object)
+    end
+  end
+end
