+++ app/serializers/v1/resource_serializer.rb
@@ -1,0 +1,13 @@
+module V1
+  class ResourceSerializer < ApplicationSerializer
+    attributes :id, :title, :excerpt, :published_at, :tags, :resource_type
+
+    def excerpt
+      object.excerpt
+    end
+
+    def tags
+      object.cached_tags
+    end
+  end
+end
